Updating Collection callback timing
I was running into an issue concerning the creation of relationships. I
suspect there is a larger issue regarding callbacks.

The switch from a blanket :before_save to two :before_validation, one
for :create and one for :update, appearred to be required. If I went
with the single :before_validation, the callbacks were fired, but the
expected behavior was incorrect.